refactor: change the directory structure - change directory name: api -> `domain` - move config directory into global: /src/main/java/com/term/fastingdatacounter/`config` -> /src/main/java/com/term/fastingdatacounter/global/`config`
