--- src/main/java/com/term/fastingdatecounter/api/food/repository/FoodRepository.java
@@ -1,10 +1,0 @@
-package com.term.fastingdatecounter.api.food.repository;
-
-import com.term.fastingdatecounter.api.food.domain.Food;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface FoodRepository extends JpaRepository<Food, Long> {
-    List<Food> findByUserId(Long userId);
-}
