--- src/main/java/com/term/fastingdatecounter/api/user/controller/dto/dto/OAuthAttributes.java
@@ -1,41 +1,0 @@
-package com.term.fastingdatecounter.api.user.controller.dto.dto;
-
-import com.term.fastingdatecounter.api.user.domain.User;
-import lombok.Builder;
-import lombok.Getter;
-
-import java.util.Map;
-
-@Getter
-public class OAuthAttributes {
-    private Map<String, Object> attributes;
-    private String nameAttributeKey;
-    private String name;
-    private String email;
-
-    @Builder
-    public OAuthAttributes(Map<String, Object> attributes, String nameAttributeKey, String name, String email) {
-        this.attributes = attributes;
-        this.nameAttributeKey = nameAttributeKey;
-        this.name = name;
-        this.email = email;
-    }
-
-    // OAuth2User에서 반환하는 사용자 정보는 Map이기 때문에 값을 변환해주어야 함
-    public static OAuthAttributes of(String userNameAttributeName, Map<String, Object> attributes){
-        return OAuthAttributes.builder()
-                .name((String) attributes.get("name"))
-                .email((String) attributes.get("email"))
-                .attributes(attributes)
-                .nameAttributeKey(userNameAttributeName)
-                .build();
-    }
-
-    // 처음 가입 시 아래 메서드가 호출되어 엔티티 생성
-    public User toEntity() {
-        return User.builder()
-                .name(name)
-                .email(email)
-                .build();
-    }
-}
