--- src/main/java/com/term/fastingdatecounter/api/user/domain/LoginUserArgumentResolver.java
@@ -1,39 +1,0 @@
-package com.term.fastingdatecounter.api.user.domain;
-
-import com.term.fastingdatecounter.api.user.controller.dto.dto.SessionUser;
-import lombok.RequiredArgsConstructor;
-import org.springframework.core.MethodParameter;
-import org.springframework.stereotype.Component;
-import org.springframework.web.bind.support.WebDataBinderFactory;
-import org.springframework.web.context.request.NativeWebRequest;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.method.support.ModelAndViewContainer;
-
-import javax.servlet.http.HttpSession;
-
-@RequiredArgsConstructor
-@Component
-public class LoginUserArgumentResolver implements HandlerMethodArgumentResolver {
-
-    private final HttpSession httpSession;
-
-    // 컨트롤러 메서드의 특정 파라미터를 지원하는지 판단
-    @Override
-    public boolean supportsParameter(MethodParameter parameter) {
-        boolean isLoginUserAnnotation = parameter.getParameterAnnotation(LoginUser.class) != null;
-        boolean isUserClass = SessionUser.class.equals(parameter.getParameterType());
-        return isLoginUserAnnotation && isUserClass; // 파라미터 앞에 @LoginUser 어노테이션 + SessionUser 클래스 타입이면 true
-    }
-
-    // 파라미터에 전달할 객체 생성
-    @Override
-    public Object resolveArgument(
-            MethodParameter parameter,
-            ModelAndViewContainer mavContainer,
-            NativeWebRequest webRequest,
-            WebDataBinderFactory binderFactory
-    ) throws Exception {
-        // 세션에서 객체 가져옴
-        return httpSession.getAttribute("user");
-    }
-}
