--- src/main/java/com/term/fastingdatecounter/api/user/domain/User.java
@@ -1,37 +1,0 @@
-package com.term.fastingdatecounter.api.user.domain;
-
-import com.term.fastingdatecounter.global.domain.BaseTimeEntity;
-import lombok.*;
-import org.hibernate.annotations.DynamicInsert;
-
-import javax.persistence.*;
-import java.io.Serializable;
-
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@DynamicInsert
-@Getter
-@Entity
-public class User extends BaseTimeEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(length = 30, nullable = false)
-    private String email;
-
-    @Column(length = 20, nullable = false)
-    private String name;
-
-    @Builder
-    public User(Long id, String email, String name) {
-        this.id = id;
-        this.email = email;
-        this.name = name;
-    }
-
-    public User update(String name){
-        this.name = name;
-        return this;
-    }
-}
