--- src/main/java/com/term/fastingdatecounter/api/user/repository/UserRepository.java
@@ -1,11 +1,0 @@
-package com.term.fastingdatecounter.api.user.repository;
-
-import com.term.fastingdatecounter.api.user.domain.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    Optional<User> findByEmail(String email);
-}
