--- src/main/java/com/term/fastingdatecounter/api/user/service/CustomOAuth2UserService.java
@@ -1,50 +1,0 @@
-package com.term.fastingdatecounter.api.user.service;
-
-import com.term.fastingdatecounter.api.user.controller.dto.dto.OAuthAttributes;
-import com.term.fastingdatecounter.api.user.controller.dto.dto.SessionUser;
-import com.term.fastingdatecounter.api.user.domain.User;
-import com.term.fastingdatecounter.api.user.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.hibernate.service.spi.ServiceException;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
-import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
-import org.springframework.security.oauth2.client.userinfo.OAuth2UserService;
-import org.springframework.security.oauth2.core.OAuth2AuthenticationException;
-import org.springframework.security.oauth2.core.user.DefaultOAuth2User;
-import org.springframework.security.oauth2.core.user.OAuth2User;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import javax.servlet.http.HttpSession;
-import java.util.Collections;
-
-
-@RequiredArgsConstructor
-@Service
-public class CustomOAuth2UserService implements OAuth2UserService<OAuth2UserRequest, OAuth2User> {
-    private final UserRepository userRepository;
-    private final HttpSession httpSession;
-
-    @Override
-    public OAuth2User loadUser(OAuth2UserRequest userRequest) throws OAuth2AuthenticationException {
-        OAuth2UserService delegate = new DefaultOAuth2UserService();
-        OAuth2User oAuth2User = delegate.loadUser(userRequest);
-
-        String userNameAttributeName = userRequest.getClientRegistration().getProviderDetails()
-                .getUserInfoEndpoint().getUserNameAttributeName(); // OAuth2 로그인 시 키가 되는 필드값(PK), 구글만 sub라는 이름으로 지원, 네이버 등의 추가 연동을 고려하여 작성
-
-        OAuthAttributes attributes = OAuthAttributes.of(userNameAttributeName, oAuth2User.getAttributes()); // OAuth2UserService를 통해 가져온 유저의 속성을 저장할 클래스
-
-        User user = saveOrUpdate(attributes);
-        httpSession.setAttribute("user", new SessionUser(user));
-        return new DefaultOAuth2User(Collections.singleton(new SimpleGrantedAuthority("ROLE_USER")), attributes.getAttributes(), attributes.getNameAttributeKey());
-    }
-
-    private User saveOrUpdate(OAuthAttributes attributes) {
-        User user = userRepository.findByEmail(attributes.getEmail())
-                .map(entity -> entity.update(attributes.getName())) // 기존 정보가 존재하면 update
-                .orElse(attributes.toEntity());
-        return userRepository.save(user);
-    }
-}
