+++ src/main/java/com/term/fastingdatecounter/domain/food/dto/FoodResponse.java
@@ -1,0 +1,26 @@
+package com.term.fastingdatecounter.domain.food.dto;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import com.term.fastingdatecounter.domain.food.domain.Food;
+import lombok.Getter;
+
+import java.time.LocalDate;
+
+@Getter
+public class FoodResponse {
+
+    private final Long id;
+    private final String name;
+
+    @JsonFormat(pattern = "yyyy-MM-dd")
+    private final LocalDate startDate;
+
+    private final Long dayCount;
+
+    public FoodResponse(Food food) {
+        this.id = food.getId();
+        this.name = food.getName();
+        this.startDate = food.getStartDate();
+        this.dayCount = food.getDayCount();
+    }
+}
