+++ src/main/java/com/term/fastingdatecounter/domain/user/domain/LoginUser.java
@@ -1,0 +1,11 @@
+package com.term.fastingdatecounter.domain.user.domain;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.PARAMETER) // 어노테이션이 생성될 수 있는 위치 - 메서드의 파라미터로 선언된 객체에서만 사용 가능
+@Retention(RetentionPolicy.RUNTIME)
+public @interface LoginUser { // 어노테이션 클래스로 지정
+}
