+++ src/main/java/com/term/fastingdatecounter/domain/user/repository/UserRepository.java
@@ -1,0 +1,11 @@
+package com.term.fastingdatecounter.domain.user.repository;
+
+import com.term.fastingdatecounter.domain.user.domain.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+
+    Optional<User> findByEmail(String email);
+}
